requirements.txt
# Data ingestion
datasets==2.21.0
pandas
tqdm
pyarrow
tiktoken

# Vector store
qdrant-client
sentence-transformers

# LLM
langchain-groq
langchain-core

# API
fastapi
uvicorn[standard]
pydantic
python-dotenv

# UI
streamlit
requests

# Utilities
loguru
